permission added

+++ apps/billmanagement/views.py
@@ -4,15 +4,18 @@ from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 from apps.billmanagement.forms import *
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import Group
 
 
 def get_signup(request):
     if request.method == "POST":
         form = UserCreationForm(request.POST)
         if form.is_valid():
-            form.save()
+            users = form.save()
+            group = Group.objects.get(name="customer")
+            users.groups.add(group)
             username = form.cleaned_data.get("username")
-            raw_password = form.cleaned_data.get("password")
+            raw_password = form.cleaned_data.get("password1")
             user = authenticate(username=username, password=raw_password)
             login(request, user)
             messages.success(request, "Signup successfully")
@@ -46,18 +49,26 @@ def get_logout(request):
 
     return redirect("signin")
 
-@login_required(login_url='signin')
-def dashboard(request):
 
-    fil = {}
-    if request.method == "POST":
-        fil["status"] = request.POST.get("status")
-    enquires = Customer.objects.filter(**fil)
-    context = {"enquires": enquires}
+@login_required(login_url="signin")
+def dashboard(request):
+    if request.user.groups.filter(permissions__name="Can view group"):
+        fil = {}
+        if request.method == "POST":
+            fil["status"] = request.POST.get("status")
+        enquires = Customer.objects.filter(**fil)
+
+    else:
+        enquires = Customer.objects.filter(user__id=request.user.id)
+
+    context = {
+        "enquires": enquires,
+    }
 
     return render(request, "pages/index.html", context)
 
-@login_required(login_url='signin')
+
+@login_required(login_url="signin")
 def create_customer(request):
 
     if request.method == "POST":
@@ -73,29 +84,41 @@ def create_customer(request):
     }
     return render(request, "pages/create_customer.html", context)
 
-@login_required(login_url='signin')
+
+@login_required(login_url="signin")
 def customer_update(request, pk):
-    if request.method == "POST":
-        customer = Customer.objects.get(id=pk)
-        form = CustomerForm(request.POST, request.FILES, instance=customer)
-        if form.is_valid():
-            form.save()
-            messages.success(request, "Customer update successfully")
+    if request.user.groups.filter(permissions__name="Can add customer"):
 
-            return redirect("dashboard")
+        if request.method == "POST":
+            customer = Customer.objects.get(id=pk)
+            form = CustomerForm(request.POST, request.FILES, instance=customer)
+            if form.is_valid():
+                form.save()
+                messages.success(request, "Customer update successfully")
+
+                return redirect("dashboard")
+        else:
+            customer = Customer.objects.get(id=pk)
+            form = CustomerForm(instance=customer)
+
+        context = {
+            "customer": customer,
+            "form": form,
+        }
     else:
-        customer = Customer.objects.get(id=pk)
-        form = CustomerForm(instance=customer)
+        messages.success(request,"your have permission to update")   
+        return redirect("dashboard")
 
-    context = {
-        "customer": customer,
-        "form": form,
-    }
     return render(request, "pages/update_customer.html", context)
 
-@login_required(login_url='signin')
+
+@login_required(login_url="signin")
 def customer_delete(request, pk):
-    customer = Customer.objects.get(id=pk)
-    customer.delete()
-    messages.success(request, "Delete successfully")
-    return redirect("/")
+    if request.user.groups.filter(permissions__name="Can delete customer"):
+        customer = Customer.objects.get(id=pk)
+        customer.delete()
+        messages.success(request, "Delete successfully")
+        return redirect("/")
+    else:
+        messages.success(request, "you dont have permission to delete")
+        return redirect("/")
